修复C类型转换编译错误。Fix C type cast compile error.

+++ internal/implementation_cgo/graphs.go
@@ -109,7 +109,7 @@ func (p *PdfiumImplementation) GetPagePath(request *requests.GetPagePath) (*resp
 			}
 
 			for j := 0; j < int(segCount); j++ {
-				seg := C.FPDFPath_GetPathSegment(obj, j)
+				seg := C.FPDFPath_GetPathSegment(obj, C.int(j))
 				segType := C.FPDFPathSegment_GetType(seg)
 				isClose := C.FPDFPathSegment_GetClose(seg)
 				var x, y C.float
